move validation to api only
